BackEnd/pyproject.toml
[tool.poetry]
name = "chat-ai-app"
version = "0.1.0"
description = "Full-stack AI chat system with child profile support"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
udantic = "^1.3.1"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
aiofiles = "^23.2.1"
jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = "^1.7.4"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
redis = "^5.0.4"
celery = "^5.4.0"
pymongo = "^4.7.2"
pydantic = "^2.7.1"

[tool.poetry.dev-dependencies]
pytest = "^8.2.1"
flake8 = "^7.0.0"
mypy = "^1.10.0"
alembic = "^1.13.1"
httpx = "^0.27.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
